+++ packages/wallet-connectors/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Schema object format conversion in `signAndSendTransaction` for WalletConnect requests.
+
 ## [0.3.1] - 2023-06-04
 
 ### Added